refactor: extract batch execution
